refactor chnaged res obj and broke out service func

+++ models/mood_tag_frequency.go
@@ -3,5 +3,5 @@ package models
 type MoodTagFrequency struct {
 	MoodTag    string  `json:"moodTag"`
 	Count      int     `json:"count"`
-	Percentage float32 `json:"percentage"`
+	Percentage float64 `json:"percentage"`
 }
